WPF control - move some of the mapping methods into extensions, they could potentially be reused
