+++ CefSharp.Wpf/CefSharp.Wpf.csproj
@@ -93,6 +93,7 @@
   <ItemGroup>
     <Compile Include="DelegateCommand.cs" />
     <Compile Include="DisposableEventWrapper.cs" />
+    <Compile Include="Internals\WpfExtensions.cs" />
     <Compile Include="Rendering\InteropBitmapInfo.cs" />
     <Compile Include="IWpfWebBrowser.cs" />
     <Compile Include="Properties\AssemblyInfo.cs">
